Support access token configuration
This is meant to replace authentication via username and password.

In combination with the new `:WriteAsAuth` command, users can
authenticate once and store all the information they need in their
.vimrc without keeping their password stored in plaintext. This keeps
their local information more secure and as a bonus makes publishing
much faster (as authentication is a slow process).

This maintains backwards-compatibility with password-based
authentication, and also uses the single _authenticate() function for
all publishing operations.